refactored spotify token to the api blueprint
